tests: Test to check NA in input (wip)
